+++ core/network/src/main/java/com/emin/network/data/MovieDataSource.kt
@@ -1,0 +1,9 @@
+package com.emin.network.data
+
+import com.emin.model.dto.MoviesDto
+
+interface MovieDataSource {
+
+    suspend fun getMovies(search : String) : MoviesDto
+
+}
