--- core/network/src/main/java/com/emin/network/data/NetworkDataSource.kt
@@ -1,4 +1,0 @@
-package com.emin.network.data
-
-class NetworkDataSource {
-}
