--- core/network/src/main/java/com/emin/network/data/NetworkDataSourceImpl.kt
@@ -1,4 +1,0 @@
-package com.emin.network.data
-
-class NetworkDataSourceImpl {
-}
