--- core/network/src/main/java/com/emin/network/di/InterceptorModule.kt
@@ -1,25 +1,0 @@
-package com.emin.network.di
-
-import com.emin.core.network.BuildConfig
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-import okhttp3.logging.HttpLoggingInterceptor
-
-@Module
-@InstallIn(SingletonComponent::class)
-object InterceptorModule {
-
-    @Provides
-    @Singleton
-    fun provideLoggingInterceptor(): HttpLoggingInterceptor {
-        return HttpLoggingInterceptor { message ->
-        }.apply {
-            if (BuildConfig.DEBUG) {
-                setLevel(HttpLoggingInterceptor.Level.BODY)
-            }
-        }
-    }
-}
