+++ core/network/src/main/java/com/emin/network/di/okhttp/InterceptorModule.kt
@@ -1,0 +1,25 @@
+package com.emin.network.di.okhttp
+
+import com.emin.core.network.BuildConfig
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+import okhttp3.logging.HttpLoggingInterceptor
+
+@Module
+@InstallIn(SingletonComponent::class)
+object InterceptorModule {
+
+    @Provides
+    @Singleton
+    fun provideLoggingInterceptor(): HttpLoggingInterceptor {
+        return HttpLoggingInterceptor { message ->
+        }.apply {
+            if (BuildConfig.DEBUG) {
+                setLevel(HttpLoggingInterceptor.Level.BODY)
+            }
+        }
+    }
+}
