+++ core/network/src/main/java/com/emin/network/di/retrofit/RetrofitModule.kt
@@ -1,0 +1,46 @@
+package com.emin.network.di.retrofit
+
+import com.emin.network.di.okhttp.OkHttpClientModule
+import com.emin.utils.Constants.BASE_URL
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import retrofit2.Converter
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module(
+    includes = [
+        OkHttpClientModule::class,
+    ],
+)
+@InstallIn(SingletonComponent::class)
+object RetrofitModule {
+
+    @Singleton
+    @Provides
+    fun provideRetrofitClient(
+        converterFactory: Converter.Factory,
+        okHttpClient: OkHttpClient,
+    ): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(converterFactory)
+            .client(okHttpClient)
+            .build()
+    }
+}
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ConverterFactoryModule {
+
+    @Singleton
+    @Provides
+    fun provideConverterFactory(): Converter.Factory {
+        return GsonConverterFactory.create()
+    }
+}
