--- core/network/src/main/java/com/emin/network/settings/AppModule.kt
@@ -1,34 +1,0 @@
-package com.emin.network.settings
-
-import com.emin.network.api.ApiService
-import com.emin.repository.MovieRepository
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AppModule {
-    private const val BASE_URL = "https://api.example.com/"
-
-    @Provides
-    @Singleton
-    fun provideMovieApi() : ApiService {
-        return Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(ApiService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideMovieRepository(api : ApiService) : MovieRepository {
-        return MovieRepositoryImpl(apiService = api)
-    }
-}
