+++ gradle/libs.versions.toml
@@ -128,6 +128,8 @@ hiltCompiler = "dagger.hilt.compiler"
 
 jetbrainsKotlinJvm = "1.9.24"
 material3Android = "1.3.1"
+runtimeAndroid = "1.7.8"
+foundationLayoutAndroid = "1.7.2"
 
 [libraries]
 
@@ -278,3 +280,5 @@ ktor-serialization = { group = "io.ktor", name = "ktor-serialization", version.r
 
 # Lottie
 lottie = { group = "com.airbnb.android", name = "lottie-compose", version.ref = "lottie" }
+androidx-runtime-android = { group = "androidx.compose.runtime", name = "runtime-android", version.ref = "runtimeAndroid" }
+androidx-foundation-layout-android = { group = "androidx.compose.foundation", name = "foundation-layout-android", version.ref = "foundationLayoutAndroid" }
