+++ presentation/settingsfeature/build.gradle.kts
@@ -29,4 +29,5 @@ dependencies {
     moduleCoreModel()
     moduleDomainUseCase()
     moduleDomainRepository()
+    modulePresentationBaseFeature()
 }
