+++ presentation/settingsfeature/src/main/java/com/emin/settingsfeature/MoviesScreen.kt
@@ -1,0 +1,118 @@
+
+import android.os.Build
+import androidx.annotation.RequiresExtension
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.emin.basefeature.MovieListRow
+import com.emin.settingsfeature.viewmodel.MoviesEvent
+import com.emin.settingsfeature.viewmodel.SettingsViewModel
+
+@RequiresExtension(extension = Build.VERSION_CODES.S, version = 7)
+@Composable
+fun SettingsScreen(viewModel: SettingsViewModel = hiltViewModel()) {
+    val state = viewModel.state.value
+
+    Box(modifier = Modifier
+        .fillMaxSize()
+        .background(Color.White)
+        .testTag("MSBox")) {
+
+        Column {
+            MovieSearchBar(modifier = Modifier
+                .fillMaxWidth()
+                .padding(20.dp),
+                hint = "Batman",
+                onSearch = {
+                    viewModel.onEvent(MoviesEvent.Search(it))
+                }
+            )
+
+            LazyColumn(modifier = Modifier.fillMaxSize().testTag("MSLazyColumn")) {
+                items(state.movies) { movie ->
+                    MovieListRow(movie = movie, onItemClick = {
+                    })
+                }
+            }
+        }
+
+        if (state.error.isNotBlank()) {
+            Text(text = state.error,
+                color = MaterialTheme.colorScheme.error,
+                textAlign = TextAlign.Center,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(14.dp)
+                    .align(Alignment.Center)
+            )
+        }
+
+        if(state.isLoading) {
+            CircularProgressIndicator(modifier = Modifier.align(Alignment.Center))
+        }
+    }
+
+}
+@Composable
+fun MovieSearchBar(
+    modifier : Modifier,
+    hint: String = "",
+    onSearch: (String) -> Unit = {}
+) {
+    var text by remember {
+        mutableStateOf("")
+    }
+    var isHintDisplayed by remember {
+        mutableStateOf(hint != "")
+    }
+
+    Box(modifier = modifier) {
+        TextField(
+            value = text,
+            keyboardActions = KeyboardActions(onDone = {
+                onSearch(text)
+            }),
+            onValueChange = {
+                text =it
+            },
+            maxLines = 1,
+            singleLine = true,
+            textStyle = TextStyle(color = Color.Black),
+            shape = RoundedCornerShape(12.dp),
+            colors = TextFieldDefaults.colors(Color.White)
+            ,
+            modifier = Modifier
+                .fillMaxWidth()
+                .shadow(5.dp, CircleShape)
+                .background(Color.White, CircleShape)
+                .padding(horizontal = 20.dp)
+                .onFocusChanged {
+                    isHintDisplayed = it.isFocused != true && text.isEmpty()
+                }
+        )
+        if(isHintDisplayed) {
+            Text(
+                text = hint,
+                color = Color.LightGray,
+                modifier = Modifier
+                    .padding(horizontal = 20.dp, vertical = 12.dp)
+            )
+        }
+    }
+}
