+++ presentation/settingsfeature/src/main/java/com/emin/settingsfeature/viewmodel/MoviesViewModel.kt
@@ -1,0 +1,64 @@
+package com.emin.settingsfeature.viewmodel
+
+import android.os.Build
+import androidx.annotation.RequiresExtension
+import androidx.compose.runtime.mutableStateOf
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.emin.common.utils.Resource
+import com.emin.model.Movie
+import com.emin.usecase.GetMoviesUseCase
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import javax.inject.Inject
+import androidx.compose.runtime.State
+
+@RequiresExtension(extension = Build.VERSION_CODES.S, version = 7)
+@HiltViewModel
+class SettingsViewModel @Inject constructor(private val getMoviesUseCase: GetMoviesUseCase) :
+    ViewModel() {
+    private val _state = mutableStateOf<MoviesState>(MoviesState())
+    val state : State<MoviesState> = _state
+
+    private var job : Job? = null
+
+    init {
+        getMovies(_state.value.search)
+    }
+
+    private fun getMovies(search: String) {
+        job?.cancel()
+        job = getMoviesUseCase.executeGetMovies(search).onEach {
+            when (it) {
+                is Resource.Success -> {
+                    _state.value = MoviesState(movies = it.data ?: emptyList())
+                }
+
+                is Resource.Error -> {
+                    _state.value = MoviesState(error = (it.error ?: "Error!").toString())
+                }
+
+                is Resource.Loading -> {
+                    _state.value = MoviesState(isLoading = true)
+                }
+            }
+        }.launchIn(viewModelScope)
+    }
+
+    fun onEvent(event : MoviesEvent) {
+        when(event) {
+            is MoviesEvent.Search -> {
+                getMovies(event.searchString)
+            }
+        }
+    }
+}
+
+data class MoviesState(
+    val isLoading : Boolean = false,
+    val movies : List<Movie> = emptyList(),
+    val error : String = "",
+    val search : String = "batman"
+)
